+++ CHANGES.rst
@@ -10,6 +10,8 @@ Unreleased (4.0.2) (XXXX-XX-XX)
 
 - Fix bug in styles for temporal rasters.
 
+- Fix raster request not including geometry id.
+
 
 Release 4.0.2 (2016-8-5)
 ---------------------